// apps/web/src/shared/api/generated/model/optionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

/**
 * 선지 정보 Response
 */
export interface OptionResponse {
  /** 선지 아이디 */
  optionId: number;
  /** 내용 */
  content: string;
  /** 설명 */
  explanation: string;
  /** 정답 여부 */
  isAnswer: boolean;
  /** 문제 아이디 */
  problemId: number;
}